[refactor] Session 기존 DB에서 분리하여 관리
