+++ src/main/java/db/Database.java
@@ -1,6 +1,5 @@
 package db;
 
-import model.Session;
 import model.User;
 
 import java.util.Collection;
@@ -11,22 +10,6 @@ import java.util.Optional;
 public class Database {
 
     private static Map<String, User> users = new HashMap<>();
-    private static Map<String, Session> sessions = new HashMap<>();
-
-    public static void addSession(Session session) {
-        sessions.put(session.getSessionId(), session);
-    }
-
-    public static Optional<Session> findSessionById(String sessionId) {
-        if(!sessions.containsKey(sessionId)){
-            return Optional.empty();
-        }
-        return Optional.ofNullable(sessions.get(sessionId));
-    }
-
-    public static void removeSession(String sessionId) {
-        sessions.remove(sessionId);
-    }
 
     public static void addUser(User user) {
         users.put(user.getUserId(), user);
